requirements.txt
asttokens==2.4.1
azure-core==1.30.1
azure-identity==1.15.0
azure-storage-blob==12.19.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
executing==2.0.1
Flask==3.0.2
Flask-Cors==4.0.0
idna==3.6
ipykernel==6.29.4
ipython==8.23.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
lxml==5.2.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
msal==1.28.0
msal-extensions==1.1.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
platformdirs==4.2.0
portalocker==2.8.2
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
threadpoolctl==3.4.0
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
